backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
supabase==2.11.0
